Try to fix deployment

+++ app/channels/session_channel.rb
@@ -9,13 +9,15 @@ class SessionChannel < ApplicationCable::Channel
   end
 
   def unsubscribed
-    Session.find_by(user_id: current_user.id).destroy
-    ActionCable.server.broadcast('session_channel', { action: 'destroy', 'valid' => true,
-                                                      online_session: current_user.username })
-    Message.where(user_id: current_user.id).each do |message|
-      message.destroy
-      ActionCable.server.broadcast('chatroom_channel', { valid: true,
-                                                         messageID: message.id, action: 'delete' })
+    if current_user
+      Session.find_by(user_id: current_user.id).destroy
+      ActionCable.server.broadcast('session_channel', { action: 'destroy', 'valid' => true,
+                                                        online_session: current_user.username })
+      Message.where(user_id: current_user.id).each do |message|
+        message.destroy
+        ActionCable.server.broadcast('chatroom_channel', { valid: true,
+                                                           messageID: message.id, action: 'delete' })
+      end
     end
   end
